+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to the "goto-package" extension will be documented in this file.
 
+## [1.0.0] - 2020-02-29
+### Added
+- Add a new command to pick package and do some actions, like go to package.json or go to npm page.
+
+## Changed
+- Remove all commands existing.
+
 ## [0.1.0] - 2019-03-09
 ### Added
 - Add "Pick and go to a package" command to pick and go to selected package.